pyproject.toml
[tool.poetry]
name = "weather-app"
version = "0.1.0"
description = "Ideal project for API studies."
authors = ["Kalebe Szlachta <you@example.com>"]
license = "Kalebe Szlachta"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
requests = "^2.32.3"
pyyaml = "^6.0.1"
flask = "^3.0.3"
Jinja2 = "^3.1.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
